% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_comparison.R
\name{full_comparison}
\alias{full_comparison}
\title{Compare the content of two rtf files}
\usage{
full_comparison(old, new)
}
\arguments{
\item{old}{character, giving the full file path and name of the old file (required)}

\item{new}{character, giving the full file path and name of the new file (required)}
}
\value{
Returns \code{diff_print}, html output of the differences
}
\description{
\code{full_comparison} returns side by side comparison of two rtf files
}
\examples{
verifyr::full_comparison(old = paste0(fs::path_package("/extdata/base_files/",
                                                          "14-1.01.rtf",
                                                          package = "verifyr")),
                         new = paste0(fs::path_package("/extdata/compare_files/",
                                                          "14-1.01.rtf",
                                                          package = "verifyr")))

}
