% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_comparison.R
\name{initial_comparison}
\alias{initial_comparison}
\title{Initial comparison of content of two files}
\usage{
initial_comparison(old, new)
}
\arguments{
\item{old}{character, giving the full file path and name of the old file (required)}

\item{new}{character, giving the full file path and name of the new file (required)}
}
\value{
Returns a vector \code{diff_print} with basic info about the differences in the content of the files
}
\description{
\code{initial_comparison} returns basic info about the differences of two files.
}
\examples{
verifyr::initial_comparison(old = "./test_outputs/base_files/14-1.01.rtf",
                            new = "./test_outputs/compare_files/14-1.01.rtf")

}
