% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List files that exist in two folders}
\usage{
list_files(old, new, pattern = NULL)
}
\arguments{
\item{old}{character, giving the the full file path and name of the folder where old files are stored (required)}

\item{new}{character, giving the the full file path and name of the folder where new files are stored (required)}

\item{pattern}{character, limit the files to be listed to contain a specific pattern (optional)}
}
\value{
Returns a tibble, \code{selected_files} with 3 columns \code{file}, \code{old_path}, \code{new_path}
}
\description{
\code{list_files} List files that exist in two folders with a specific name pattern
}
\examples{
verifyr::list_files(old = "./test_outputs/base_files",
                    new = "./test_outputs/compare_files",
                    pattern = "14-1")

}
