% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\title{Call for shiny example where the user can test verifyr package functions}
\usage{
run_example()
}
\description{
\code{run_example} returns simple Shiny App where user can see how the verifyr functions work
}
\details{
@examples
verifyr::run_example()
}
