% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Call for shiny example where the user can test verifyr package functions}
\usage{
runExample()
}
\description{
\code{runExample} returns simple Shiny App where user can see how the verifyr functions work
}
\details{
@examples
verifyr::runExample()
}
